requirements.txt
flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0